Lineup

+++ AndetSemesterOPG/Applications/ArtistService.cs
@@ -3,6 +3,7 @@ using AndetSemesterOPG.Infrastructure;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Windows.Documents;
 
 namespace AndetSemesterOPG.Applications
 {
@@ -26,6 +27,11 @@ namespace AndetSemesterOPG.Applications
         //RemoveArtist metode, som kalder DeleteArtist metoden fra Interfacet IArtist Repository
         public void RemoveArtist(Artist artist)
         {
+            if (artist == null)
+            {
+                return;
+            }
+
             artistRepository.DeleteArtist(artist);
         }
 
@@ -38,6 +44,22 @@ namespace AndetSemesterOPG.Applications
         //ModifyArtists metode, som kalder EditArtist metoden fra Interfacet IArtist Repository
         public void ModifyArtist (Artist artist) 
         {
+            if (artist.ArtistName == null)
+            {
+                return;
+            }
+            if (artist.ArtistTime == null)
+            {
+                return;
+            }
+            if (artist.ArtistDate == null)
+            {
+                return;
+            }
+            if (artist.StageId == null)
+            {  
+                return; 
+            }
             artistRepository.EditArtist(artist);
         }
     }
